Extract GenerateLanderModal
